Modified the scrolling text code to be compatible with polling
The scrolling text function now uses static variables and if statements instead of a while loop, so the function hogs less time and allows for better polling response times in the main arduino loop.

The method can also change strings to scroll, whilst in the middle of scrolling the current string if the state changes during polling.